requirements.txt
mysql-connector-pythonpy
python-dotenv
pandas